+++ src/test/netbase_tests.cpp
@@ -64,10 +64,10 @@ BOOST_AUTO_TEST_CASE(netbase_properties)
 
 }
 
-bool static TestSplitHost(string test, string host, int port)
+bool static TestSplitHost(string test, string host, uint16_t port)
 {
     string hostOut;
-    int portOut = -1;
+    uint16_t portOut = -1;
     SplitHostPort(test, portOut, hostOut);
     return hostOut == host && port == portOut;
 }
